runtime cast correction, threads.h include path fix, and removal of leftover debug logging

+++ runtime/c11/threads_pthreads.c
@@ -40,7 +40,7 @@ int thrd_join(thrd_t thr, int *res) {
     }
     
     if (res) {
-        *res = (int)code;
+        *res = (intptr_t)code;
     }
     
     return thrd_success;
